move box_id to ergo_box;
